slack_bot_prototype/requirements.txt
# Core dependencies for Slack Community Bot
slack-sdk>=3.23.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0